+++ src/Cp/Components/ComponentRegistry.php
@@ -37,6 +37,7 @@ class ComponentRegistry
         'lightswitch' => Lightswitch::class,
         'money-input' => MoneyInput::class,
         'number-input' => NumberInput::class,
+        'object-select' => ObjectSelect::class,
         'option-rows' => OptionRows::class,
         'radio' => Radio::class,
         'radio-group' => RadioGroup::class,